data-ingest/pyproject.toml
[tool.poetry]
name = "data-ingest"
version = "0.1.0"
description = ""
authors = ["João Correia <joaodpscorreia@gmail.com>"]
packages = [{include = "data_ingest"}]

[tool.poetry.dependencies]
python = "~3.9"
pandas = "^2.1.4"
yfinance = "^0.2.33"
hopsworks = "^3.7.0"
great-expectations = "0.18.12"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = "^6.2.5"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["E501"]
select = ["B","C","E","F","W","B9", "I", "Q"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

[tool.ruff.mccabe]
max-complexity = 10
